+++ application/classes/Core/Theme.php
@@ -16,7 +16,7 @@ namespace Core {
 		* Doesn't do much of interest, this shouldn't be required to mess with.
 		* @return void
 		*/
-		public function __construct($themename) {
+		public function __construct(string $themename) {
 			$this->name = $themename;
 			$this->theme = (new Configuration($this->getTemplatePath("/theme.json")));
 
@@ -27,7 +27,7 @@ namespace Core {
 		* Get the path to the current active theme.
 		* @return string
 		*/
-		public function getTemplatePath($tpl = '') {
+		public function getTemplatePath(string $tpl = '') : string {
 			$path = \Registry::get("Core\Application")->getApplicationPath()."/application/themes/".$this->getName();
 
 			if($tpl != '') {
@@ -44,7 +44,7 @@ namespace Core {
 		* @author Allan Thue Rehhoff
 		* @return string
 		*/
-		public function getDirectoryUri($file = '/') {
+		public function getDirectoryUri(string $file = '/') : string {
 			$protocol = SSL ? "https://" : "http://";
 			$host  = isset($_SERVER["SERVER_NAME"]) ? $_SERVER["SERVER_NAME"] : "127.0.0.1";
 			$path = rtrim(dirname($_SERVER["PHP_SELF"]), "/\\");
@@ -57,7 +57,7 @@ namespace Core {
 		* Get the current theme name loaded.
 		* @return string
 		*/
-		public function getName() {
+		public function getName() : string {
 			return $this->name;
 		}
 
@@ -68,7 +68,7 @@ namespace Core {
 		* @uses \Document
 		* @return void
 		*/
-		private function addAssets() {
+		private function addAssets() : void {
 			if(!empty($this->theme->get("javascript"))) {
 				foreach($this->theme->get("javascript") as $javascript) {
 					if(filter_var($javascript, FILTER_VALIDATE_URL)) {