+++ preprocess.php
@@ -22,7 +22,7 @@
 
 	// Output buffering
 	// The ob_gzhandler callback returns false if browser doesn't support gzip
-	!CLI ? ob_start("ob_gzhandler") ? : ob_start() : false;
+	//!CLI ? ob_start("ob_gzhandler") ? : ob_start() : false;
 
 	// Error reporting
 	ini_set("display_errors", "On");